requirements.txt
# DST Submittals Generator - Requirements
python-docx>=0.8.11
pypdf>=3.0.0
docx2pdf>=0.1.8
reportlab>=4.0.0
pywin32>=306